+++ src/design/plone/contenttypes/behaviors/additional_help_infos.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from collective import dexteritytextindexer
+from plone.app.dexterity import textindexer
 from design.plone.contenttypes import _
 from collective.volto.blocksfield.field import BlocksField
 from plone.autoform.interfaces import IFormFieldProvider
@@ -29,7 +29,7 @@ class IAdditionalHelpInfos(model.Schema):
         fields=["ulteriori_informazioni"],
     )
 
-    dexteritytextindexer.searchable("ulteriori_informazioni")
+    textindexer.searchable("ulteriori_informazioni")
 
 
 @implementer(IAdditionalHelpInfos)