+++ src/design/plone/contenttypes/behaviors/news_additional_fields.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from collective import dexteritytextindexer
+from plone.app.dexterity import textindexer
 from collective.volto.blocksfield.field import BlocksField
 from design.plone.contenttypes import _
 from plone.app.z3cform.widget import RelatedItemsFieldWidget
@@ -133,7 +133,7 @@ class INewsAdditionalFields(model.Schema):
     form.order_before(numero_progressivo_cs="ILeadImageBehavior.image")
     form.order_before(a_cura_di="ILeadImageBehavior.image")
 
-    dexteritytextindexer.searchable("descrizione_estesa")
+    textindexer.searchable("descrizione_estesa")
 
 
 @implementer(INewsAdditionalFields)