+++ src/design/plone/contenttypes/indexers/pagina_argomento.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-from collective import dexteritytextindexer
 from design.plone.contenttypes.interfaces.pagina_argomento import IPaginaArgomento
+from plone.app.dexterity.textindexer.interfaces import IDynamicTextIndexExtender
+from plone.restapi.indexers import SearchableText_blocks
 from zope.component import adapter
 from zope.interface import implementer
-from plone.restapi.indexers import SearchableText_blocks
 
 
 @adapter(IPaginaArgomento)
-@implementer(dexteritytextindexer.IDynamicTextIndexExtender)
+@implementer(IDynamicTextIndexExtender)
 class SearchableTextExtender(object):
     def __init__(self, context):
         self.context = context